feat: unit components preview in sidebar
