+++ src/library-authoring/components/ContainerCard.tsx
@@ -55,7 +55,7 @@ type ContainerCardProps = {
 
 const ContainerCard = ({ hit } : ContainerCardProps) => {
   const { componentPickerMode } = useComponentPickerContext();
-  const { showOnlyPublished } = useLibraryContext();
+  const { setUnitId, showOnlyPublished } = useLibraryContext();
   const { openUnitInfoSidebar } = useSidebarContext();
 
   const {
@@ -81,7 +81,7 @@ const ContainerCard = ({ hit } : ContainerCardProps) => {
   const openContainer = useCallback(() => {
     if (itemType === 'unit') {
       openUnitInfoSidebar(unitId);
-
+      setUnitId(unitId);
       navigateTo({ unitId });
     }
   }, [unitId, itemType, openUnitInfoSidebar, navigateTo]);
